fix(cli-sub-agent): honor tier bypass gate for exact model specs
Address #1749 pre-PR review round 2 findings.\n\n- thread the tier bypass gate decision into exact --model-spec resolution\n- preserve default tier whitelist rejection when bypass is not allowed\n- order preferred tier candidates by preference-array priority\n- split MCP csa_run handling below the source token gate
