+++ crates/cli-sub-agent/src/debate_cmd_resolve.rs
@@ -128,6 +128,11 @@ pub(crate) fn resolve_debate_selection(
                 needs_edit: false,
                 tier: cli_tier,
                 force_ignore_tier_setting,
+                model_spec_tier_bypass_allowed: crate::run_helpers::tier_bypass_allowed(
+                    project_config,
+                    global_config,
+                    false,
+                ),
                 tool_is_auto_resolved: false,
             })?;
         return Ok(ResolvedDebateSelection {
